перенёс кнопку переключения темы в отдельный файл вместе с тестами
