+++ src/widgets/header/header.module.scss
@@ -9,18 +9,6 @@
   background-color: var(--color-interactive-brand-hover);
 }
 
-// Очень хочу добавиить, но, возможно, будет лагать
-// *,
-// *::before,
-// *::after {
-//   transition:
-//     background-color 0.3s,
-//     color 0.3s,
-//     border-color 0.3s,
-//     fill 0.3s,
-//     stroke 0.3s;
-// }
-
 @keyframes theme-icon-enter {
   from {
     transform: rotate(-90deg) scale(0.5);
